+++ R/modifiers.r
@@ -45,6 +45,10 @@ NULL
 #' @export
 #' @rdname modifiers
 fixed <- function(pattern, ignore_case = FALSE) {
+  if (!is_bare_character(pattern)) {
+    stop("Can only modify plain character vectors.", call. = FALSE)
+  }
+
   options <- stri_opts_fixed(case_insensitive = ignore_case)
 
   structure(
@@ -63,6 +67,10 @@ fixed <- function(pattern, ignore_case = FALSE) {
 #'   \code{\link[stringi]{stri_opts_regex}}, or
 #'   \code{\link[stringi]{stri_opts_brkiter}}
 coll <- function(pattern, ignore_case = FALSE, locale = NULL, ...) {
+  if (!is_bare_character(pattern)) {
+    stop("Can only modify plain character vectors.", call. = FALSE)
+  }
+
   options <- stri_opts_collator(
     strength = if (ignore_case) 2L else 3L,
     locale = locale,
@@ -86,6 +94,10 @@ coll <- function(pattern, ignore_case = FALSE, locale = NULL, ...) {
 #' @param dotall If \code{TRUE}, \code{.} will also match line terminators.
 regex <- function(pattern, ignore_case = FALSE, multiline = FALSE,
                    comments = FALSE, dotall = FALSE, ...) {
+  if (!is_bare_character(pattern)) {
+    stop("Can only modify plain character vectors.", call. = FALSE)
+  }
+
   options <- stri_opts_regex(
     case_insensitive = ignore_case,
     multiline = multiline,
@@ -149,4 +161,8 @@ ignore.case <- function(string) {
 perl <- function(pattern) {
   message("perl is deprecated. Please use regex() instead")
   regex(pattern)
+}
+
+is_bare_character <- function(x) {
+  is.character(x) && !is.object(x)
 }